+++ src/bunit.web/JSInterop/InvocationHandlers/JSRuntimeInvocationHandlerBase{TResult}.cs
@@ -8,11 +8,6 @@ namespace Bunit.JSInterop.InvocationHandlers
 	/// </summary>
 	public abstract class JSRuntimeInvocationHandlerBase<TResult>
 	{
-		/// <summary>
-		/// The identifier string used to indicate a catch all handler.
-		/// </summary>
-		protected internal const string CatchAllIdentifier = "*";
-
 		private readonly InvocationMatcher invocationMatcher;
 		private TaskCompletionSource<TResult> completionSource;
 
@@ -22,14 +17,9 @@ namespace Bunit.JSInterop.InvocationHandlers
 		public virtual bool IsVoidResultHandler { get; }
 
 		/// <summary>
-		/// Gets a value indicating whether this handler will match any invocations that expect a <typeparamref name="TResult"/> as the return type.
+		/// Gets a value indicating whether this handler is considered a catch all handler for invocations with <typeparamref name="TResult"/> as the return type.
 		/// </summary>
 		public bool IsCatchAllHandler { get; }
-
-		/// <summary>
-		/// Gets the expected identifier for the function to invoke.
-		/// </summary>
-		public string Identifier { get; }
 
 		/// <summary>
 		/// Gets the invocations that this <see cref="JSRuntimeInvocationHandler{TResult}"/> has matched with.
@@ -39,17 +29,13 @@ namespace Bunit.JSInterop.InvocationHandlers
 		/// <summary>
 		/// Initializes a new instance of the <see cref="JSRuntimeInvocationHandlerBase{TResult}"/> class.
 		/// </summary>
-		/// <param name="identifier">Identifier it matches. Set to "*" to match all identifiers.</param>
 		/// <param name="matcher">An invocation matcher used to determine if the handler should handle an invocation.</param>
-		protected JSRuntimeInvocationHandlerBase(string identifier, InvocationMatcher matcher)
+		/// <param name="isCatchAllHandler">Set to true if this handler is a catch all handler, that should only be used if there are no other non-catch all handlers available.</param>
+		protected JSRuntimeInvocationHandlerBase(InvocationMatcher matcher, bool isCatchAllHandler)
 		{
-			if (string.IsNullOrWhiteSpace(identifier))
-				throw new ArgumentException($"'{nameof(identifier)}' cannot be null or whitespace", nameof(identifier));
-
-			Identifier = identifier;
-			IsCatchAllHandler = string.Equals(identifier, CatchAllIdentifier, StringComparison.Ordinal);
 			invocationMatcher = matcher ?? throw new ArgumentNullException(nameof(matcher));
 			completionSource = new TaskCompletionSource<TResult>();
+			IsCatchAllHandler = isCatchAllHandler;
 		}
 
 		/// <summary>
@@ -107,8 +93,6 @@ namespace Bunit.JSInterop.InvocationHandlers
 		/// </summary>
 		/// <param name="invocation">Invocation to check.</param>
 		/// <returns>True if the handler can handle the invocation, false otherwise.</returns>
-		internal bool CanHandle(JSRuntimeInvocation invocation) => (IsCatchAllHandler || MatchesIdentifier(invocation)) && invocationMatcher(invocation);
-
-		private bool MatchesIdentifier(JSRuntimeInvocation invocation) => Identifier.Equals(invocation.Identifier, StringComparison.Ordinal);
+		internal bool CanHandle(JSRuntimeInvocation invocation) => invocationMatcher(invocation);
 	}
 }